Musekontrol til kortest vej
Tilføjet metode i Tiler-klassen, som bør tegne en glass-pane med den
nuværende koteste vej. (OBS: translateToView-metoden virker ikke,
hvorfor tegning af den korteste vej endnu ikke virker.)
Tilføjet musekontrol til navigation - højreklik for startpunkt og
venstreklik for slutpunkt. Der udregnes et kortest søgetræ fra punktet
ved højreklik mens der ved vestreklik forespørges den korteste vej fra
start- til slutpunktet.
Tilføjet linjebredde til paintLines-metoden i Painter-klassen.

+++ Map/Line.java
@@ -8,9 +8,9 @@ public class Line {
 
 	public Vector start, stop;
 	public Color color;
-	public double width;
+	public float width;
 
-	public Line set(Vector start, Vector stop, Color color, double width) {
+	public Line set(Vector start, Vector stop, Color color, float width) {
 		this.start = start;
 		this.stop = stop;
 		this.color = color;
